+++ ReleaseNotes.md
@@ -1,6 +1,7 @@
 ﻿# Release notes of Ampersand
 
 ## unreleased
+- [Issue #1576](https://github.com/AmpersandTarski/Ampersand/issues/1576) When importing .ttl files, only warn if a label is missing. 
 - [Issue #1578](https://github.com/AmpersandTarski/Ampersand/issues/1578) Harvest PATTERNs based on skos:ConceptScheme 
 
 ## v5.4.6 to v5.4.8